Move GitRatchet into lib-extra so it can be used by plugin-maven.
