+++ lib-extra/src/main/java/com/diffplug/spotless/extra/GitRatchet.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.gradle.spotless;
+package com.diffplug.spotless.extra;
 
 import java.io.File;
 import java.io.IOException;
