+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/GradleGitRatchet.java
@@ -21,6 +21,8 @@ import javax.annotation.Nullable;
 
 import org.gradle.api.Project;
 
+import com.diffplug.spotless.extra.GitRatchet;
+
 /** Gradle implementation of GitRatchet. */
 class GradleGitRatchet extends GitRatchet<Project> {
 	@Override
